fix: #48 - PR 코멘트 반영
- View extension과 내부 struct 분리
- 일부 ViewModel 밖으로 나와있던 변수를 ViewModel로 합침
- 잘못 작성된 파일명 수정
    - 회원가입(이메일 입력 View 이름)
    - 커스텀 네비게이션 백 버튼 네이밍 수정
