+++ Honzapda/Presentation/Onboarding/View/Main/SignUpPasswordCheckTypeView.swift
@@ -11,9 +11,9 @@ import SwiftUI
 struct SignUpPasswordCheckTypeView: View {
     // MARK: PARAMS
     @Environment(\.dismiss) private var dismiss
-    @ObservedObject var signUpViewModel = SignUpViewModel()
+    @ObservedObject var signUpViewModel: SignUpViewModel
     @State private var inputText = ""
-    let placeholder = "비밀번호 확인"
+    private let placeholder = "비밀번호 확인"
     @State private var isNextButtonClicked = false
     
     // MARK: BODY
@@ -66,12 +66,8 @@ struct SignUpPasswordCheckTypeView: View {
             transaction.disablesAnimations = true
         })
         .navigationBarBackButtonHidden(true)
-        .navigationBarItems(leading: customNavigationBackButton(color: .black,
+        .navigationBarItems(leading: CustomNavigationBackButton(color: .black,
                                                                 title: "",
                                                                 dismiss: self.dismiss))
     } //: BODY
 } //: 회원가입 비밀번호 확인 입력 View
-
-#Preview {
-    SignUpPasswordCheckTypeView()
-}
